Replaced admin login back to interceptor
