+++ board-game-statistics-api/src/main/java/com/asd/board_game_statistics_api/config/WebConfig.java
@@ -1,11 +1,18 @@
 package com.asd.board_game_statistics_api.config;
 
+import com.asd.board_game_statistics_api.admin.AdminAuthenticationInterceptor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.lang.NonNull;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
+
+    @Value("${BGS_ADMIN_PASSWORD}")
+    private String adminPassword;
 
     @Override
     public void addCorsMappings(CorsRegistry registry) {
@@ -14,6 +21,12 @@ public class WebConfig implements WebMvcConfigurer {
                 .allowedMethods("GET", "POST", "PUT", "DELETE")
                 .allowedHeaders("*")
                 .allowCredentials(true);
+    }
+
+    @Override
+    public void addInterceptors(@NonNull InterceptorRegistry registry) {
+        registry.addInterceptor(new AdminAuthenticationInterceptor(adminPassword))
+                .addPathPatterns("/api/admin/**");
     }
 
 }
